requirements.txt
black==22.3.0
coverage==6.4.1
dj-database-url==0.5.0
Django==3.2.9
django-bootstrap4==22.1
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-heroku==0.3.1
djangorestframework==3.12.4
docutils==0.18
gunicorn==20.1.0
Markdown==3.3.6
Pillow==9.2.0
psycopg==3.0.9
psycopg2==2.9.3
pytest-django==4.5.2
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
